requirements.txt
Django
djangorestframework
drf-spectacular
drf-spectacular-sidecar
psycopg2-binary
